Added drag and drop

+++ controllers/columnsController.js
@@ -1,4 +1,7 @@
 const Column = require('../models/column');
+// ===-VR-===
+const Board = require('../models/board');
+// ===-VR-===
 const { controllerWrapper } = require('../decorators');
 const { HttpError } = require('../helpers');
 const Task = require('../models/task');
@@ -26,8 +29,17 @@ const getAllColumnsWithTasks = async (req, res) => {
 };
 
 const addColumn = async (req, res) => {
+  // ===-VR-===
+  const { parentBoard } = req.body;
+  // ===-VR-===
   const result = await Column.create({ ...req.body });
-
+  // ===-VR-===
+  await Board.findByIdAndUpdate(
+    parentBoard,
+    { $push: { columnOrder: result._id } },
+    { safe: true, upsert: true, new: true }
+  );
+  // ===-VR-===
   res.status(201).json(result);
 };
 
@@ -45,7 +57,15 @@ const editColumn = async (req, res) => {
 
 const deleteColumn = async (req, res) => {
   const { id } = req.params;
-
+  // ===-VR-===
+  const column = await Column.findById(id);
+  if (!column) {
+    throw HttpError(404, 'Column not found');
+  }
+  await Board.findByIdAndUpdate(column.parentBoard, {
+    $pull: { columnOrder: column._id },
+  });
+  // ===-VR-===
   const result = await Column.findByIdAndRemove(id);
   if (!result) {
     throw HttpError(404, 'Column not found');
@@ -60,9 +80,23 @@ const deleteColumn = async (req, res) => {
   res.status(204).json({ message: `Column with id: ${id} deleted` });
 };
 
+// ===-VR-===
+const editTaskOrder = async (req, res) => {
+  const { id } = req.params;
+  const result = await Column.findByIdAndUpdate(id, req.body, { new: true });
+  if (!result) {
+    throw HttpError(404, `Column with id: ${id} not found`);
+  }
+  res.status(201).json(result);
+};
+// ===-VR-===
+
 module.exports = {
   getAllColumnsWithTasks: controllerWrapper(getAllColumnsWithTasks),
   addColumn: controllerWrapper(addColumn),
   editColumn: controllerWrapper(editColumn),
   deleteColumn: controllerWrapper(deleteColumn),
+  // ===-VR-===
+  editTaskOrder: controllerWrapper(editTaskOrder),
+  // ===-VR-===
 };
